refactor: move processor creation into own file
